+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.2] - 2018-01-29
 ### Changed
 
 - Mangafox domain from `http://mangafox.la` to `http://fanfox.net`.
@@ -110,3 +112,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [0.4.0-beta]: https://github.com/enakada/foxy-manga-reader/compare/v0.3.1...v0.4.0-beta
 [0.5.0]: https://github.com/enakada/foxy-manga-reader/compare/v0.4.0-beta...v0.5.0
 [0.5.1]: https://github.com/enakada/foxy-manga-reader/compare/v0.5.0...v0.5.1
+[0.5.2]: https://github.com/enakada/foxy-manga-reader/compare/v0.5.1...v0.5.2